requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
Django==5.0
djangorestframework==3.14.0
filelock==3.13.1
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
platformdirs==3.11.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tzdata==2023.3
virtualenv==20.24.6
